redgrid/pyproject.toml
[project]
name = "redgrid"
version = "0.1.0"
description = "OpenEnv environment for traffic signal optimization via RL"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "setuptools",
    "wheel",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"
